+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Enhancements
   * [Suggestions] Upgrading elixir_phoenix suggestion to work with Phoenix v1.0.0
 
+* Enhancements
+  * [Docker] Checking Docker version (azk supports Docker v1.8+);
+
 ## v0.14.6 - (2015-08-20)
 
 * Bug